requirements.txt
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.2.0
langchain>=0.1.0
tqdm>=4.65.0
tenacity>=8.2.0
fastapi>=0.95.0
uvicorn>=0.22.0
json5>=0.9.14
pydantic>=2.0.0
typing-extensions>=4.7.0
networkx>=3.1.0
pygraphviz>=1.10.0
